+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableFloatArray.kt
@@ -532,6 +532,15 @@ public value class ImmutableFloatArray @PublishedApi internal constructor(
         return values.toSet().toImmutableArray()
     }
 
+    /**
+     * Returns an immutable array containing only the elements having distinct keys returned by the
+     * [selector]
+     */
+    public inline fun <K> distinctBy(selector: (element: Float) -> K): ImmutableFloatArray {
+        val keys = HashSet<K>()
+        return filter { keys.add(selector(it)) }
+    }
+
     public companion object {
         @PublishedApi
         internal val EMPTY: ImmutableFloatArray = ImmutableFloatArray(FloatArray(0))