+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableShortArray.kt
@@ -533,6 +533,15 @@ public value class ImmutableShortArray @PublishedApi internal constructor(
         return values.toSet().toImmutableArray()
     }
 
+    /**
+     * Returns an immutable array containing only the elements having distinct keys returned by the
+     * [selector]
+     */
+    public inline fun <K> distinctBy(selector: (element: Short) -> K): ImmutableShortArray {
+        val keys = HashSet<K>()
+        return filter { keys.add(selector(it)) }
+    }
+
     public companion object {
         @PublishedApi
         internal val EMPTY: ImmutableShortArray = ImmutableShortArray(ShortArray(0))